Split StringRect and StringDraw
