+++ Makefile.am
@@ -71,7 +71,8 @@ lib_explorer_la_SOURCES = $(explorer)/window_explorer.cpp \
 													$(explorer)/image_window_explorer.cpp \
 													$(explorer)/controller.cpp \
 													$(explorer)/image_cell_explorer.cpp \
-													$(explorer)/string_draw.cpp
+													$(explorer)/string_draw.cpp \
+													$(explorer)/string_rect.cpp
 
 lib_explorer_la_LIBADD = lib_handler.la \
 						 lib_thread.la \
