+++ src/explorer/string_rect.cpp
@@ -1,0 +1,40 @@
+/*
+ *  Copyright (c) 2021 Uskrai
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#include "fmr/explorer/string_rect.h"
+
+#include <wx/dc.h>
+
+#include "fmr/nowide/string.h"
+
+namespace fmr {
+
+namespace explorer {
+void StringRect::SetString(const std::string &string) {
+  string_ = String::Widen<wxString>(string);
+}
+
+std::string StringRect::GetString() const { return String::Narrow(string_); }
+
+void StringRect::Draw(wxDC &dc) {
+  dc.DrawText(string_, GetRect().GetPosition());
+  ;
+}
+
+}  // namespace explorer
+
+}  // namespace fmr
